it works but lacks light/dark switch
