+++ app/components/data-display.js
