+++ app/lib/report.js
