+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "nuoya-capital-website",
   "description": "nuoya.capital",
-  "version": "1.1.0",
+  "version": "1.2.0",
   "author": "Bokang Huang <hbk671104@gmail.com>",
   "private": true,
   "scripts": {
@@ -13,19 +13,20 @@
   },
   "dependencies": {
     "@chakra-ui/icons": "^2.1.0",
+    "@chakra-ui/next-js": "^2.1.5",
     "@chakra-ui/react": "^2.8.0",
     "@emotion/react": "^11.11.1",
     "@emotion/styled": "^11.11.0",
     "framer-motion": "^6.5.1",
     "got": "^12.6.1",
-    "next": "12.1.6",
-    "react": "18.1.0",
-    "react-dom": "18.1.0"
+    "next": "^13.4.19",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
     "autoprefixer": "^10.4.14",
     "eslint": "7.32.0",
-    "eslint-config-next": "12.1.6",
+    "eslint-config-next": "^13.4.19",
     "postcss": "^8.4.27",
     "tailwindcss": "^3.3.3"
   }
