--- pages/_app.js
@@ -1,14 +1,0 @@
-import '../styles/globals.css'
-import { ChakraProvider } from '@chakra-ui/react'
-
-import theme from '../styles/theme'
-
-function MyApp({ Component, pageProps }) {
-  return (
-    <ChakraProvider theme={theme}>
-      <Component {...pageProps} />
-    </ChakraProvider>
-  )
-}
-
-export default MyApp
