--- pages/_document.js
@@ -1,18 +1,0 @@
-import { Html, Head, Main, NextScript } from 'next/document'
-import { ColorModeScript } from '@chakra-ui/react'
-
-import theme from '../styles/theme'
-
-export default function Document() {
-  return (
-    <Html lang="en">
-      <Head />
-      <body>
-        {/* 👇 Here's the script */}
-        <ColorModeScript initialColorMode={theme.config.initialColorMode} />
-        <Main />
-        <NextScript />
-      </body>
-    </Html>
-  )
-}
